Move profile field elements into their own files
